+++ ui/src/app/cohort-review/query-cohort-definition/query-cohort-definition.component.html
@@ -1,0 +1,26 @@
+<!--Cohort Definition HTML start-->
+<div class="query-title">
+  Cohort Definition
+</div>
+<ng-container *ngFor="let group of definition; let g = index;">
+  <ng-container *ngFor="let item of group?.groups; let i = index">
+    <div *ngIf="i > 0">AND</div>
+    <div class="group-background">
+      <ng-container *ngFor="let param of item; let p = index;">
+        <div *ngIf="p > 0 ">OR</div>
+        <div style="border: 1px solid black;">
+          <ng-container *ngFor="let query of param; let s = index;">
+            <ng-container *ngIf="s > 0">
+              <div *ngIf="query?.type ==='DEMO'; else or ">AND</div>
+              <ng-template #or>OR</ng-template>
+            </ng-container>
+          <div>
+          {{query?.items}}
+          </div>
+          </ng-container>
+        </div>
+      </ng-container>
+    </div>
+  </ng-container>
+</ng-container>
+<!--Cohort Definition HTML END-->
