+++ ui/src/app/cohort-review/query-report/query-report.component.html
@@ -1,97 +1,144 @@
-<div class="container doNotPrint">
-  <div class="row header-border">
-    <div class="col-sm-1 col-xs-1 btn-margin">
-      <button
-        type="button"
-        class="btn btn-sm"
-        title="Go Back to the Participant Table"
-        [routerLink]="['../participants']">
-        <clr-icon shape="angle-double" dir="left"></clr-icon>
-        Participants
-      </button>
-    </div>
-      <h4 class="col-sm-11 col-xs-11">Cohort Query Report</h4>
-  </div>
-</div>
-
 <div class="report-background">
   <div class="container">
     <div class="row">
-      <div class="col-sm-10 col-xs-10">
-        Cohort name: {{cohort.name}} <br>
-        Created by: {{cohort.creator}} <br>
-        Date created: {{ cohort.creationTime | date: 'yyyy-MM-dd': 'UTC/GMT'}} <br><br>
-      </div>
-      <div class="col-sm-2 col-xs-2">
-        <clr-icon class="doNotPrint" shape="printer" size="48"></clr-icon>
-        <button class="doNotPrint btn btn-primary" (click)="onPrint()">PRINT REPORT</button>
-      </div>
-    </div>
-  </div>
-
-<!--Cohort Definition HTML start-->
-  Cohort Definition: <br>
-  <ng-container *ngFor="let group of definition; let g = index;">
-    <ng-container *ngFor="let item of group?.groups; let i = index">
-     <div *ngIf="i > 0">AND</div>
-      <div class="group-background">
-        <ng-container *ngFor="let param of item; let p = index;">
-          <div *ngIf="p > 0 ">OR</div>
-          <div style="border: 1px solid black;">
-            <ng-container *ngFor="let query of param; let s = index;">
-              <ng-container *ngIf="s > 0">
-                <div *ngIf="query?.type ==='DEMO'; else or ">AND</div>
-                <ng-template #or>OR</ng-template>
-              </ng-container>
-            <div>
-              {{query?.items}}
+      <div class="col-sm-6 col-xs-6">
+        <div class="container">
+          <div class="row">
+            <div class="col-sm-6 col-xs-6">
+              <div class="query-title">
+                Cohort Name
+              </div>
+              <div class="query-content">
+                {{cohort.name}}
+              </div>
+              <div class="query-title">
+                Created By
+              </div>
+              <div class="query-content">
+                {{cohort.creator}}
+              </div>
             </div>
-            </ng-container>
+            <div class="col-sm-6 col-xs-6">
+              <div class="query-title">
+                Date created
+              </div>
+              <div class="query-content">
+                {{cohort.creationTime | date: 'yyyy-MM-dd': 'UTC/GMT'}}
+              </div>
+              <div class="query-title">
+                CDR Version
+              </div>
+              <div class="query-content">
+                {{cdrId}}
+              </div>
+            </div>
+            <div class="col-sm-12 col-xs-12">
+              <app-query-cohort-definition [cohort]="cohort" [review]="review"></app-query-cohort-definition>
+            </div>
           </div>
-        </ng-container>
+        </div>
       </div>
-    </ng-container>
-  </ng-container>
-  <!--Cohort Definition HTML END-->
+      <div class="col-sm-6">
 
-  <!--Descriptive HTML Start-->
-  <div>
-    Descriptive Statistics:<br>
-    <div style="width: 60%; border: 1px solid black;">
-      <div class="container">
-        <div class="row">
-          <div class="col-sm-4"></div>
-          <div class="col-sm-4">
-            Total Count
-          </div>
-          <div class="col-sm-4">
-            Percent of Cohort
-          </div>
-        </div>
-        <div class="row">
-          <div class="col-sm-12">
-              Gender
-          </div>
-        </div>
-        <div class="row">
-          <div class="col-sm-4">
-            <div>Male</div>
-          </div>
-          <div class="col-sm-4">
-               1027
-          </div>
-          <div class="col-sm-4">
-            100%
+        <!--Descriptive HTML Start-->
+        <div>
+          Descriptive Statistics:<br>
+          <div style="width: 60%; border: 1px solid black;">
+            <div class="container">
+              <div class="row">
+                <div class="col-sm-4"></div>
+                <div class="col-sm-4">
+                Total Count
+                </div>
+                <div class="col-sm-4">
+                Percent of Cohort
+                </div>
+              </div>
+              <div class="row">
+                <div class="col-sm-12">
+                Gender
+                </div>
+              </div>
+              <div class="row">
+                <div class="col-sm-4">
+                  <div>Male</div>
+                </div>
+                <div class="col-sm-4">
+                  1027
+                </div>
+                <div class="col-sm-4">
+                  100%
+                </div>
+              </div>
+            </div>
           </div>
         </div>
 
+        <!--Descriptive HTML END-->
       </div>
     </div>
   </div>
-
-  <!--Descriptive HTML END-->
-    <app-overview-charts class="test-print"></app-overview-charts>
+  <app-overview-charts class="test-print"></app-overview-charts>
 </div>
+
+<!--<div class="container doNotPrint">-->
+  <!--<div class="row header-border">-->
+    <!--<div class="col-sm-1 col-xs-1 btn-margin">-->
+      <!--<button-->
+        <!--type="button"-->
+        <!--class="btn btn-sm"-->
+        <!--title="Go Back to the Participant Table"-->
+        <!--[routerLink]="['../participants']">-->
+        <!--<clr-icon shape="angle-double" dir="left"></clr-icon>-->
+        <!--Participants-->
+      <!--</button>-->
+    <!--</div>-->
+      <!--<h4 class="col-sm-11 col-xs-11">Cohort Query Report</h4>-->
+  <!--</div>-->
+<!--</div>-->
+
+<!--<div class="report-background">-->
+  <!--<div class="container">-->
+    <!--<div class="row">-->
+      <!--<div class="col-sm-10 col-xs-10">-->
+        <!--Cohort name: {{cohort.name}} <br>-->
+        <!--Created by: {{cohort.creator}} <br>-->
+        <!--Date created: {{ cohort.creationTime | date: 'yyyy-MM-dd': 'UTC/GMT'}} <br><br>-->
+      <!--</div>-->
+      <!--<div class="col-sm-2 col-xs-2">-->
+        <!--<clr-icon class="doNotPrint" shape="printer" size="48"></clr-icon>-->
+        <!--<button class="doNotPrint btn btn-primary" (click)="onPrint()">PRINT REPORT</button>-->
+      <!--</div>-->
+    <!--</div>-->
+  <!--</div>-->
+
+<!--&lt;!&ndash;Cohort Definition HTML start&ndash;&gt;-->
+  <!--Cohort Definition: <br>-->
+  <!--<ng-container *ngFor="let group of definition; let g = index;">-->
+    <!--<ng-container *ngFor="let item of group?.groups; let i = index">-->
+     <!--<div *ngIf="i > 0">AND</div>-->
+      <!--<div class="group-background">-->
+        <!--<ng-container *ngFor="let param of item; let p = index;">-->
+          <!--<div *ngIf="p > 0 ">OR</div>-->
+          <!--<div style="border: 1px solid black;">-->
+            <!--<ng-container *ngFor="let query of param; let s = index;">-->
+              <!--<ng-container *ngIf="s > 0">-->
+                <!--<div *ngIf="query?.type ==='DEMO'; else or ">AND</div>-->
+                <!--<ng-template #or>OR</ng-template>-->
+              <!--</ng-container>-->
+            <!--<div>-->
+              <!--{{query?.items}}-->
+            <!--</div>-->
+            <!--</ng-container>-->
+          <!--</div>-->
+        <!--</ng-container>-->
+      <!--</div>-->
+    <!--</ng-container>-->
+  <!--</ng-container>-->
+  <!--&lt;!&ndash;Cohort Definition HTML END&ndash;&gt;-->
+
+
+<!--</div>-->
